pushsync: optimize TestPushsyncSimulation
